# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_MarlonG1_api-facturacion-sv_internal_domain_auth_models.AuthCredentials:
    properties:
      api_key:
        type: string
      api_secret:
        type: string
      credentials:
        $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_domain_auth_models.HaciendaCredentials'
    type: object
  github_com_MarlonG1_api-facturacion-sv_internal_domain_auth_models.HaciendaCredentials:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  github_com_MarlonG1_api-facturacion-sv_internal_domain_core_user.Address:
    properties:
      complement:
        type: string
      department:
        type: string
      municipality:
        type: string
    type: object
  github_com_MarlonG1_api-facturacion-sv_internal_domain_core_user.BranchOffice:
    properties:
      address:
        $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_domain_core_user.Address'
      api_key:
        type: string
      api_secret:
        type: string
      email:
        type: string
      establishment_code:
        type: string
      establishment_code_mh:
        type: string
      establishment_type:
        type: string
      is_active:
        type: boolean
      phone:
        type: string
      pos_code:
        type: string
      pos_code_mh:
        type: string
      user:
        $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_domain_core_user.User'
    type: object
  github_com_MarlonG1_api-facturacion-sv_internal_domain_core_user.ListBranchesResponse:
    properties:
      api_key:
        type: string
      api_secret:
        type: string
      branch_number:
        type: integer
      establishment_code:
        type: string
      establishment_type:
        type: string
    type: object
  github_com_MarlonG1_api-facturacion-sv_internal_domain_core_user.User:
    properties:
      auth_type:
        type: string
      branch_offices:
        description: Relationships
        items:
          $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_domain_core_user.BranchOffice'
        type: array
      business_name:
        type: string
      commercial_name:
        type: string
      economic_activity:
        type: string
      economic_activity_desc:
        type: string
      email:
        type: string
      nit:
        type: string
      nrc:
        type: string
      password_pri:
        type: string
      phone:
        type: string
      token_lifetime:
        type: integer
      year_in_dte:
        type: boolean
    type: object
  github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError:
    properties:
      code:
        type: string
      details:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
host: 185.222.241.94:7319
info:
  contact:
    email: soporte@empresa.com
    name: Soporte API DTE
    url: https://github.com/CorenaDeveloper/api-dte-CrediExpress
  description: API para la gestión de Documentos Tributarios Electrónicos (DTE) que
    cumple con los requisitos establecidos por la autoridad fiscal de El Salvador.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: API Facturación El Salvador
  version: "1.0"
paths:
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_domain_core_user.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_domain_core_user.ListBranchesResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
      summary: Register
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with API Key, API Secret and Hacienda Credentials
      parameters:
      - description: Auth credentials
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_domain_auth_models.AuthCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
      security:
      - BearerAuth: []
      summary: Login
      tags:
      - Auth
  /dte:
    get:
      consumes:
      - application/json
      description: Obtiene lista paginada de DTEs del usuario autenticado
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - default: 1
        description: Número de página
        in: query
        name: page
        type: integer
      - default: 10
        description: Elementos por página
        in: query
        name: limit
        type: integer
      - description: Filtrar por estado
        enum:
        - PROCESSED
        - PENDING
        - ERROR
        in: query
        name: status
        type: string
      - description: Filtrar por tipo DTE
        enum:
        - "01"
        - "03"
        - "05"
        - "06"
        - "11"
        - "14"
        in: query
        name: dte_type
        type: string
      - description: Fecha inicio (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: Fecha fin (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
      security:
      - BearerAuth: []
      summary: Listar DTEs
      tags:
      - DTE
  /dte/{id}:
    get:
      consumes:
      - application/json
      description: Obtiene un DTE específico por su código de generación
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Código de generación del DTE
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
      security:
      - BearerAuth: []
      summary: Obtener DTE por código de generación
      tags:
      - DTE
  /dte/ccf:
    post:
      consumes:
      - application/json
      description: Crea un nuevo CCF electrónico DTE tipo 03
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Datos del CCF
        in: body
        name: ccf
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
      security:
      - BearerAuth: []
      summary: Crear Comprobante de Crédito Fiscal
      tags:
      - DTE
  /dte/creditnote:
    post:
      consumes:
      - application/json
      description: Crea una nueva Nota de Crédito electrónica DTE tipo 05
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Datos de la nota de crédito
        in: body
        name: creditnote
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
      security:
      - BearerAuth: []
      summary: Crear Nota de Crédito
      tags:
      - DTE
  /dte/invalidation:
    post:
      consumes:
      - application/json
      description: Invalida un DTE existente
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Datos para invalidación
        in: body
        name: invalidation
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
      security:
      - BearerAuth: []
      summary: Invalidar DTE
      tags:
      - DTE
  /dte/invoices:
    post:
      consumes:
      - application/json
      description: Crea una nueva factura electrónica DTE tipo 01
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Datos de la factura
        in: body
        name: invoice
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
      security:
      - BearerAuth: []
      summary: Crear factura electrónica
      tags:
      - DTE
  /dte/retention:
    post:
      consumes:
      - application/json
      description: Crea un nuevo Comprobante de Retención electrónico DTE tipo 06
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Datos del comprobante de retención
        in: body
        name: retention
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
      security:
      - BearerAuth: []
      summary: Crear Comprobante de Retención
      tags:
      - DTE
  /health:
    get:
      consumes:
      - application/json
      description: Check the health of all core service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
      summary: Health Check
      tags:
      - Health
  /metrics:
    get:
      consumes:
      - application/json
      description: Get metrics for a specific endpoint
      parameters:
      - description: Token JWT with Format 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Endpoint to filter metrics
        in: query
        name: endpoint
        type: string
      - description: HTTP method to filter metrics
        in: query
        name: method
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
      security:
      - BearerAuth: []
      summary: Get endpoint metrics
      tags:
      - Metrics
  /test:
    get:
      consumes:
      - application/json
      description: Run system test
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MarlonG1_api-facturacion-sv_internal_infrastructure_api_response.APIError'
      summary: Run system test
      tags:
      - Test
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"